requirements.txt
absl-py==1.4.0
appdirs==1.4.3
apturl==0.5.2
asgiref==3.7.2
astunparse==1.6.3
autopep8==2.0.2
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cachetools==5.3.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==7.1.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
contourpy==1.0.7
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.2.1
djangorestframework==3.14.0
dlib==19.24.1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
flatbuffers==23.5.9
fonttools==4.39.0
future==0.18.2
gast==0.4.0
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
gunicorn==20.1.0
h5py==3.8.0
httplib2==0.14.0
idna==2.8
importlib-metadata==6.6.0
importlib-resources==5.12.0
imutils==0.5.4
itsdangerous==1.1.0
jax==0.4.10
Jinja2==2.11.3
joblib==1.2.0
keras==2.12.0
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==16.0.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.3
MarkupSafe==1.1.1
matplotlib==3.7.1
ml-dtypes==0.1.0
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.20.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==1.2.4
paramiko==2.6.0
pexpect==4.6.0
Pillow==8.2.0
playsound==1.3.0
protobuf==4.23.1
psutil==5.5.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.16.2
pycodestyle==2.10.0
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
pytz==2021.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.15.0
sklearn==0.0.post1
sqlparse==0.4.4
systemd-python==234
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.8.1+cpu
torchvision==0.9.1+cpu
typing-extensions==4.6.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.0.17
wadllib==1.3.3
Werkzeug==1.0.1
WoeUSB-ng==0.2.12
wrapt==1.14.1
wxPython==4.0.7
xkit==0.0.0
zipp==3.15.0
